+++ src/main/java/hexlet/code/service/UserServiceImpl.java
@@ -3,10 +3,12 @@ package hexlet.code.service;
 import hexlet.code.common.mapper.UserMapper;
 import hexlet.code.dto.CreateUserDto;
 import hexlet.code.dto.GetUserDto;
+import hexlet.code.entity.Task;
 import hexlet.code.entity.User;
 import hexlet.code.exception.ExceptionMessage;
-import hexlet.code.exception.ResourceNotFoundException;
+import hexlet.code.exception.IllegalOperationException;
 import hexlet.code.exception.ResourceExistException;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -19,12 +21,17 @@ import java.util.stream.Collectors;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final TaskRepository taskRepository;
     private final UserMapper mapper;
     private final PasswordEncoder encoder;
 
     @Autowired
-    public UserServiceImpl(UserRepository userRepository, UserMapper mapper, PasswordEncoder encoder) {
+    public UserServiceImpl(UserRepository userRepository,
+                           TaskRepository taskRepository,
+                           UserMapper mapper,
+                           PasswordEncoder encoder) {
         this.userRepository = userRepository;
+        this.taskRepository = taskRepository;
         this.mapper = mapper;
         this.encoder = encoder;
     }
@@ -69,8 +76,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long id) {
-        User user = userRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(ExceptionMessage.USER_NOT_FOUND));
+        User user = userRepository.getById(id);
+        List<Task> authorTask = taskRepository.findAllByAuthorId(user.getId());
+        List<Task> executorTask = taskRepository.findAllByExecutorId(user.getId());
+        if (!authorTask.isEmpty() || !executorTask.isEmpty()) {
+            throw new IllegalOperationException(ExceptionMessage.ILLEGAL_DELETE_USER);
+        }
         userRepository.delete(user);
     }
 }
